+++ app/routes/inventory.py
@@ -40,7 +40,7 @@ def delete_product(product_id: int, db: Session = Depends(database.get_db)):
 from fastapi import Body
 
 @router.put("/products/{product_name}")
-def update_product_quantity(product_name: str, quantity: int = Body(..., embed=True), db: Session = Depends(database.get_db)):
+def update_product_quantity(product_name: str, quantity: float = Body(..., embed=True), db: Session = Depends(database.get_db)):
     # Capitalize product name before searching/adding
     product_name_cap = product_name.strip().capitalize()
     product = crud.get_product_by_name(db, product_name_cap)
